charts/requirements.txt
fastapi
uvicorn
matplotlib
scipy
pandas
python-multipart
minio
python-dotenv